mcp-server/requirements.txt
chromadb==0.6.3
sentence-transformers==3.4.1
openai==1.63.2
flask==3.1.0
flask-cors==5.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.36
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.0